fix file detection for cljs
after spending way to much time playing around inside the cljs compiler the final solution is pretty easy.
this handles the precompiled case + compiling a file in the repl. tested it in calva with shadow-cljs. 
sending forms to the repl from a file only returns "repl-input.cljs", but that might be a calva thing.