Prefer inspector to vega viewers
Since implementing a correct and complete spec for vega/vega-lite is
difficult, the spec was loosened up. However, now data that is not vega
is being matched for the vega viewer. By preferring the inspector, the
vega spec can remain open while providing a more reasonable default for
most data.